fix(network): map missing-interface delete to not-found in the client
The rollback helper and Delete paths relied on a string match because PVE answers a 400 parameter error for a missing interface, which the shared matcher only maps for 404 and 500. Translate it once in DeleteNetworkInterface so callers use errors.Is.

Move the rollback acceptance tests into their resource test files and the helpers into helpers_test.go, and only revert the node's network config in cleanup when the shadow file carries no diff.
