+++ fwprovider/nodes/network/helpers_test.go
@@ -13,13 +13,24 @@ package network_test
 
 import (
 	"context"
+	"crypto/tls"
+	"encoding/json"
 	"fmt"
+	"io"
 	"net/http"
+	"net/http/httptest"
+	"net/http/httputil"
+	"net/url"
+	"strconv"
+	"strings"
+	"testing"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
+	"github.com/stretchr/testify/require"
 
 	"github.com/bpg/terraform-provider-proxmox/fwprovider/test"
+	"github.com/bpg/terraform-provider-proxmox/utils"
 )
 
 // checkInterfaceActive asserts whether PVE reports an interface as active on the node.
@@ -64,22 +75,30 @@ func reloadNode(te *test.Environment) resource.TestCheckFunc {
 // from an applied one. PVE only sends `changes` while something is pending.
 func checkNodePendingChanges(te *test.Environment, want bool) resource.TestCheckFunc {
 	return func(_ *terraform.State) error {
-		var body struct {
-			Changes *string `json:"changes"`
+		got, err := nodeHasPendingChanges(te)
+		if err != nil {
+			return err
 		}
 
-		path := fmt.Sprintf("nodes/%s/network", te.NodeName)
-		if err := te.Client().DoRequest(context.Background(), http.MethodGet, path, nil, &body); err != nil {
-			return fmt.Errorf("reading node network: %w", err)
-		}
-
-		got := body.Changes != nil && *body.Changes != ""
 		if got != want {
 			return fmt.Errorf("node %q: pending changes = %t, want %t", te.NodeName, got, want)
 		}
 
 		return nil
 	}
+}
+
+func nodeHasPendingChanges(te *test.Environment) (bool, error) {
+	var body struct {
+		Changes *string `json:"changes"`
+	}
+
+	path := fmt.Sprintf("nodes/%s/network", te.NodeName)
+	if err := te.Client().DoRequest(context.Background(), http.MethodGet, path, nil, &body); err != nil {
+		return false, fmt.Errorf("reading node network: %w", err)
+	}
+
+	return body.Changes != nil && *body.Changes != "", nil
 }
 
 // checkStagedDestroy asserts that a destroy with `reload = false` left its delete staged rather
@@ -89,4 +108,126 @@ func checkStagedDestroy(te *test.Environment) resource.TestCheckFunc {
 		checkNodePendingChanges(te, true),
 		reloadNode(te),
 	)
+}
+
+// newInterfaceListDropProxy fronts the real PVE endpoint and removes the named interface from every node
+// interface list response, which is what a privilege-separated token without SDN.Audit sees for bridges.
+func newInterfaceListDropProxy(t *testing.T, iface string) string {
+	t.Helper()
+
+	target, err := url.Parse(utils.GetAnyStringEnv("PROXMOX_VE_ENDPOINT"))
+	require.NoError(t, err)
+
+	proxy := &httputil.ReverseProxy{
+		Rewrite: func(r *httputil.ProxyRequest) {
+			r.SetURL(target)
+			r.Out.Header.Del("Accept-Encoding")
+		},
+		Transport: &http.Transport{
+			Proxy:           http.ProxyFromEnvironment,
+			TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
+		},
+		ModifyResponse: func(resp *http.Response) error {
+			if resp.Request.Method != http.MethodGet || !strings.HasSuffix(resp.Request.URL.Path, "/network") {
+				return nil
+			}
+
+			raw, err := io.ReadAll(resp.Body)
+			_ = resp.Body.Close()
+
+			if err != nil {
+				return err
+			}
+
+			var payload struct {
+				Data []map[string]any `json:"data"`
+			}
+
+			if err := json.Unmarshal(raw, &payload); err != nil {
+				return err
+			}
+
+			kept := make([]map[string]any, 0, len(payload.Data))
+
+			for _, entry := range payload.Data {
+				if entry["iface"] != iface {
+					kept = append(kept, entry)
+				}
+			}
+
+			body, err := json.Marshal(map[string]any{"data": kept})
+			if err != nil {
+				return err
+			}
+
+			resp.Body = io.NopCloser(strings.NewReader(string(body)))
+			resp.ContentLength = int64(len(body))
+			resp.Header.Set("Content-Length", strconv.Itoa(len(body)))
+
+			return nil
+		},
+	}
+
+	server := httptest.NewTLSServer(proxy)
+	t.Cleanup(server.Close)
+
+	return server.URL
+}
+
+// cleanupStagedInterface removes the interface through the real endpoint if a failed create left it staged, then
+// clears the shadow file. The revert is node-wide, so it only runs when the shadow carries no diff and there is
+// nothing of anyone else's to discard.
+func cleanupStagedInterface(t *testing.T, te *test.Environment, iface string) {
+	t.Helper()
+
+	t.Cleanup(func() {
+		ctx := context.Background()
+
+		ifaces, err := te.NodeClient().ListNetworkInterfaces(ctx)
+		if err != nil {
+			t.Logf("cleanup: list interfaces: %v", err)
+
+			return
+		}
+
+		for _, i := range ifaces {
+			if i.Iface != iface {
+				continue
+			}
+
+			if err := te.NodeClient().DeleteNetworkInterface(ctx, iface); err != nil {
+				t.Logf("cleanup: delete interface %s: %v", iface, err)
+			}
+		}
+
+		pending, err := nodeHasPendingChanges(te)
+		if err != nil {
+			t.Logf("cleanup: %v", err)
+
+			return
+		}
+
+		if pending {
+			t.Logf("cleanup: node %s has unrelated pending network changes, leaving them", te.NodeName)
+
+			return
+		}
+
+		if err := te.NodeClient().RevertNetworkConfiguration(ctx); err != nil {
+			t.Logf("cleanup: revert network configuration: %v", err)
+		}
+	})
+}
+
+// requireInterfaceNotStaged lists interfaces through the real endpoint, which includes pending entries. It runs
+// after resource.Test because the harness skips CheckDestroy when the final state is empty.
+func requireInterfaceNotStaged(t *testing.T, te *test.Environment, iface string) {
+	t.Helper()
+
+	ifaces, err := te.NodeClient().ListNetworkInterfaces(context.Background())
+	require.NoError(t, err)
+
+	for _, i := range ifaces {
+		require.NotEqual(t, iface, i.Iface, "interface %s is still staged on node %s after the failed create", iface, te.NodeName)
+	}
 }
