+++ fwprovider/nodes/network/resource_linux_bond_test.go
@@ -14,6 +14,7 @@ package network_test
 import (
 	"fmt"
 	"os"
+	"regexp"
 	"testing"
 
 	"github.com/brianvoe/gofakeit/v7"
@@ -290,4 +291,39 @@ func TestAccResourceLinuxBondStagedDestroy(t *testing.T) {
 			},
 		},
 	})
+}
+
+func TestAccResourceLinuxBondCreateRollsBackOnReadMiss(t *testing.T) {
+	te := test.InitEnvironment(t)
+
+	slave1 := os.Getenv("PROXMOX_VE_ACC_BOND_SLAVE1")
+	slave2 := os.Getenv("PROXMOX_VE_ACC_BOND_SLAVE2")
+
+	if slave1 == "" || slave2 == "" {
+		t.Skip("skipping: PROXMOX_VE_ACC_BOND_SLAVE1 and PROXMOX_VE_ACC_BOND_SLAVE2 must be set to eth-type interfaces")
+	}
+
+	iface := fmt.Sprintf("bond%d", gofakeit.Number(10, 9999))
+	cleanupStagedInterface(t, te, iface)
+
+	endpoint := newInterfaceListDropProxy(t, iface)
+
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: te.AccProviders,
+		Steps: []resource.TestStep{
+			{
+				Config: te.RenderConfig(fmt.Sprintf(`
+				resource "proxmox_network_linux_bond" "test_rollback" {
+					name           = "%s"
+					node_name      = "{{.NodeName}}"
+					slaves         = ["%s", "%s"]
+					timeout_reload = 60
+				}
+				`, iface, slave1, slave2), test.WithInsecureEndpoint(endpoint)),
+				ExpectError: regexp.MustCompile(`Unable to Read Linux Bond After Creation`),
+			},
+		},
+	})
+
+	requireInterfaceNotStaged(t, te, iface)
 }
